pyproject.toml
[tool.poetry]
name = "spai"
version = "0.1.0"
description = "Search, Parse, and Iterate - A tool for structured data extraction from search results"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "spai"}]

[tool.poetry.dependencies]
python = "^3.9"
google-api-python-client = "^2.0.0"
google-cloud-aiplatform = "^1.38.1"
vertexai = "^0.0.1"
aiohttp = "^3.9.1"
polars = "^0.20.3"
rich = "^13.7.0"
httpx = "^0.26.0"
pydantic = "^2.5.3"
typer = "^0.9.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
spai = "spai.cli:main"
